+++ core/data/src/main/java/com/wap/wapp/core/data/repository/auth/AuthRepository.kt
@@ -6,4 +6,6 @@ interface AuthRepository {
     suspend fun deleteUser(): Result<Unit>
 
     suspend fun isUserSignIn(): Result<Boolean>
+
+    suspend fun validationWapCode(code: String): Result<Boolean>
 }
