--- core/data/src/main/java/com/wap/wapp/core/data/repository/auth/SignUpRepository.kt
@@ -1,5 +1,0 @@
-package com.wap.wapp.core.data.repository.auth
-
-interface SignUpRepository {
-    suspend fun validationWapCode(code: String): Result<Boolean>
-}
