+++ core/network/src/main/java/com/wap/wapp/core/network/source/auth/AuthDataSource.kt
@@ -6,4 +6,6 @@ interface AuthDataSource {
     suspend fun deleteUser(): Result<Unit>
 
     suspend fun isUserSignIn(): Result<Boolean>
+
+    suspend fun validationWapCode(code: String): Result<Boolean>
 }
