--- core/network/src/main/java/com/wap/wapp/core/network/source/auth/SignUpDataSource.kt
@@ -1,5 +1,0 @@
-package com.wap.wapp.core.network.source.auth
-
-interface SignUpDataSource {
-    suspend fun validationWapCode(code: String): Result<Boolean>
-}
